Refactor mastering defaults, dtypes, and EQ logic
Several interrelated changes to audio mastering and DSP internals:

- Config defaults updated: target_lufs -> -11.0, smoothing_fraction defaults to None, drive_db -> 0.0 and new config params added for bass/treble: stop_bass_boost_hz, start_treble_boost_hz, bass_boost_db_per_oct, treble_boost_ratio.
- DSP processing precision increased: chunk processing now uses float64 for input, results and window sums to improve numerical accuracy.
- Exciter adjustments: removed clipping on computed spectral gains and simplified drive calculation to remove an outer clip.
- Filters: freq_cut now forces float32 input.
- AudioEqualizer / SmartMastering changes:
  - Reduced FIR/tap sizes (taps default 4095, FIR trims to 4096) and adjusted kaiser window beta to 18.0.
  - Anchor normalization simplified: ensure low/high endpoints, work with raw freqs/gains and use nyquist = sr/2 - 0.1.
  - Dense-count logic simplified and capped differently.
  - SmartMastering now computes anchors from new bass/treble config fields and derives low_cut/high_cut defaults; target spectrum computation simplified to a flat pass-through (ones) for now.
  - Smoothing of target_db only applied when smoothing_fraction is set; removed prior phase-centering/clipping steps and aggressive phase-limiting logic.
- Minor: small I/O/save_audio whitespace and test updates (removed two tests that depended on the old spectrum/normalization behavior).

These changes aim to simplify and stabilize numeric behavior, change default mastering voicing, and reduce heavy filter kernel sizes for performance. Adjust downstream logic accordingly as some legacy assumptions (slope-based target spectrum, phase normalization) were removed.

+++ src/definers/audio/dsp.py
@@ -514,14 +514,14 @@ def process_audio_chunks(fn, data, chunk_size, overlap=0):
     if overlap >= chunk_size:
         raise ValueError("Overlap must be smaller than chunk size")
 
-    data = data.astype(np.float32)
+    data = data.astype(np.float64)
     if data.ndim == 1:
         data = data[np.newaxis, :]
 
     (_num_channels, audio_length) = data.shape
     step = chunk_size - overlap
-    final_result = np.zeros_like(data, dtype=np.float32)
-    window_sum = np.zeros_like(data, dtype=np.float32)
+    final_result = np.zeros_like(data, dtype=np.float64)
+    window_sum = np.zeros_like(data, dtype=np.float64)
     window = np.hanning(chunk_size)
     window = window[np.newaxis, :]
